+++ Automation/ScriptingEngine/ScriptableComponents/Components/WorkplaceScriptableComponent.cs
@@ -3,10 +3,12 @@
 // License: Public Domain
 
 using System;
+using System.Collections.Generic;
 using IgorZ.Automation.AutomationSystem;
 using IgorZ.Automation.ScriptingEngine.Core;
 using IgorZ.Automation.ScriptingEngine.Expressions;
 using Timberborn.BaseComponentSystem;
+using Timberborn.PrioritySystem;
 using Timberborn.WorkSystem;
 
 namespace IgorZ.Automation.ScriptingEngine.ScriptableComponents.Components;
@@ -15,9 +17,13 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
 
   const string RemoveWorkersActionLocKey = "IgorZ.Automation.Scriptable.Workplace.Action.RemoveWorkers";
   const string SetWorkersActionLocKey = "IgorZ.Automation.Scriptable.Workplace.Action.SetWorkers";
+  const string SetPriorityActionLocKey = "IgorZ.Automation.Scriptable.Workplace.Action.SetPriority";
+  const string AssignedWorkersSignalLocKey = "IgorZ.Automation.Scriptable.Workplace.Signal.AssignedWorkers";
 
   const string RemoveWorkersActionName = "Workplace.RemoveWorkers";
   const string SetWorkersActionName = "Workplace.SetWorkers";
+  const string SetPriorityActionName = "Workplace.SetPriority";
+  const string AssignedWorkersSignalName = "Workplace.AssignedWorkers";
 
   #region ScriptableComponentBase implementation
 
@@ -25,14 +31,73 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
   public override string Name => "Workplace";
 
   /// <inheritdoc/>
+  public override string[] GetSignalNamesForBuilding(AutomationBehavior behavior) {
+    var workplace = GetWorkplace(behavior, throwIfNotFound: false);
+    return workplace ? [AssignedWorkersSignalName] : [];
+  }
+
+  /// <inheritdoc/>
+  public override Func<ScriptValue> GetSignalSource(string name, AutomationBehavior behavior) {
+    var workplace = GetWorkplace(behavior);
+    return name switch {
+        AssignedWorkersSignalName => () => ScriptValue.FromInt(workplace.NumberOfAssignedWorkers),
+        _ => throw new UnknownSignalException(name),
+    };
+  }
+
+  /// <inheritdoc/>
+  public override SignalDef GetSignalDefinition(string name, AutomationBehavior behavior) {
+    var workplace = GetWorkplace(behavior);
+    return name switch {
+        AssignedWorkersSignalName => LookupSignalDef(
+            AssignedWorkersSignalName + "-" + workplace.MaxWorkers,
+            () => MakeAssignedWorkersSignalDef(workplace)),
+        _ => throw new UnknownSignalException(name),
+    };
+  }
+
+  /// <inheritdoc/>
+  public override void RegisterSignalChangeCallback(SignalOperator signalOperator, ISignalListener host) {
+    var name = signalOperator.SignalName;
+    if (name == AssignedWorkersSignalName) {
+      host.Behavior.GetOrCreate<WorkplaceChangeTracker>().AddSignal(signalOperator, host);
+    } else {
+      throw new InvalidOperationException("Unknown signal: " + name);
+    }
+  }
+
+  /// <inheritdoc/>
+  public override void UnregisterSignalChangeCallback(SignalOperator signalOperator, ISignalListener host) {
+    var name = signalOperator.SignalName;
+    if (name== AssignedWorkersSignalName) {
+      host.Behavior.GetOrThrow<WorkplaceChangeTracker>().RemoveSignal(signalOperator, host);
+    } else {
+      throw new InvalidOperationException("Unknown signal: " + name);
+    }
+  }
+
+  /// <inheritdoc/>
   public override string[] GetActionNamesForBuilding(AutomationBehavior behavior) {
     var workplace = GetWorkplace(behavior, throwIfNotFound: false);
-    return workplace ? [RemoveWorkersActionName, SetWorkersActionName] : [];
+    if (!workplace) {
+      return [];
+    }
+    var workplacePriority = behavior.GetComponent<WorkplacePriority>();
+    return workplacePriority
+        ? [RemoveWorkersActionName, SetWorkersActionName, SetPriorityActionName]
+        : [RemoveWorkersActionName, SetWorkersActionName];
   }
 
   /// <inheritdoc/>
   public override Action<ScriptValue[]> GetActionExecutor(string name, AutomationBehavior behavior) {
     var workplace = GetWorkplace(behavior);
+    if (name == SetPriorityActionName) {
+      var priorityComponent = behavior.GetComponent<WorkplacePriority>();
+      if (!priorityComponent) {
+        throw new ScriptError.BadStateError(behavior, "Building doesn't have WorkplacePriority");
+      }
+      return args => SetPriorityAction(priorityComponent, args);
+    }
     return name switch {
         RemoveWorkersActionName => _ => ResetWorkersAction(workplace),
         SetWorkersActionName => args => SetWorkersAction(workplace, args),
@@ -47,7 +112,25 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
     return name switch {
         RemoveWorkersActionName => RemoveWorkersActionDef,
         SetWorkersActionName => LookupActionDef(key, () => MakeSetWorkersActionDef(workplace)),
+        SetPriorityActionName => SetPriorityActionDef,
         _ => throw new UnknownActionException(name),
+    };
+  }
+
+  #endregion
+
+  #region Signals
+
+  SignalDef MakeAssignedWorkersSignalDef(Workplace workplace) {
+    return new SignalDef {
+        ScriptName = AssignedWorkersSignalName,
+        DisplayName = Loc.T(AssignedWorkersSignalLocKey),
+        Result = new ValueDef {
+            ValueType = ScriptValue.TypeEnum.Number,
+            ValueFormatter = x => x.AsFloat.ToString("0"),
+            ValueValidator = ValueDef.RangeCheckValidatorInt(0, workplace.MaxWorkers),
+            ValueUiHint = GetArgumentMinMaxValueHint(0, workplace.MaxWorkers),
+        },
     };
   }
 
@@ -77,6 +160,24 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
     };
   }
 
+  ActionDef SetPriorityActionDef => _setPriorityActionDef ??= new ActionDef {
+      ScriptName = SetPriorityActionName,
+      DisplayName = Loc.T(SetPriorityActionLocKey),
+      Arguments = [
+          new ValueDef {
+              ValueType = ScriptValue.TypeEnum.String,
+              Options = [
+                  ("VeryLow", Loc.T("Priorities.VeryLow")),
+                  ("Low", Loc.T("Priorities.Low")),
+                  ("Normal", Loc.T("Priorities.Normal")),
+                  ("High", Loc.T("Priorities.High")),
+                  ("VeryHigh", Loc.T("Priorities.VeryHigh")),
+              ],
+          },
+      ],
+  };
+  ActionDef _setPriorityActionDef;
+
   static void ResetWorkersAction(Workplace building) {
     building.DesiredWorkers = 0;
     building.UnassignWorkerIfOverstaffed();
@@ -95,6 +196,18 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
     building.UnassignWorkerIfOverstaffed();
   }
 
+  static void SetPriorityAction(WorkplacePriority workplacePriority, ScriptValue[] args) {
+    AssertActionArgsCount(SetPriorityActionName, args, 1);
+    var priorityName = args[0].AsString;
+    if (!Enum.TryParse<Priority>(priorityName, out var priority)) {
+      throw new ScriptError.ValueOutOfRange($"Unknown priority: {priorityName}");
+    }
+    if (workplacePriority.Priority == priority) {
+      return;
+    }
+    workplacePriority.SetPriority(priority);
+  }
+
   #endregion
 
   #region Implementation
@@ -105,6 +218,25 @@ sealed class WorkplaceScriptableComponent : ScriptableComponentBase {
       throw new ScriptError.BadStateError(building, "Building doesn't have Workplace");
     }
     return workplace;
+  }
+
+  #endregion
+
+  #region Workplace change tracker
+
+  internal sealed class WorkplaceChangeTracker : AbstractStatusTracker {
+
+    /// <inheritdoc/>
+    public override void Start() {
+      base.Start();
+      var workplace = AutomationBehavior.GetComponent<Workplace>();
+      workplace.WorkerAssigned += OnWorkerChanged;
+      workplace.WorkerUnassigned += OnWorkerChanged;
+    }
+
+    void OnWorkerChanged(object sender, WorkerChangedEventArgs args) {
+      ScheduleSignal(AssignedWorkersSignalName, ignoreErrors: true);
+    }
   }
 
   #endregion